MGNLUI-2667 Introduced DataSourceManager interface, JCR-based implementation, added field to content app descriptor.
MGNLUI-2666 Extracted WorkbenchPresenterBase and AbstractContentPresenterBase which are JCR-agnostic bases of the WorkbenchComponent, got rid of Strings as Item ids in interfaces.

MGNLUI-2667 Further development of DataSourceManager concept - moved it from app to sub-app scope in order to be able to actually control the containers of ContentPresenters, integrated DSManager further into Workbench and ContentPresenter concept, managed to get rid of WB extension classes

MGNLUI-2667 Applied changes to the configured descriptor classes.